Migrate backup admin page header to unified AdminHeader component. (#47368)
