+++ projects/packages/backup/changelog/update-normalize-backup-admin-header
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+Migrate admin page header to use unified AdminHeader component from jetpack-components.
