+++ projects/packages/backup/src/js/components/Admin/index.js
@@ -11,18 +11,27 @@ import { useConnectionErrorNotice, ConnectionError } from '@automattic/jetpack-c
 import { shouldUseInternalLinks } from '@automattic/jetpack-shared-extension-utils';
 import apiFetch from '@wordpress/api-fetch';
 import { ExternalLink } from '@wordpress/components';
-import { createInterpolateElement, useState, useEffect, useCallback } from '@wordpress/element';
+import { useSelect } from '@wordpress/data';
+import {
+	createInterpolateElement,
+	useState,
+	useEffect,
+	useCallback,
+	useMemo,
+} from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import useAnalytics from '../../hooks/useAnalytics';
 import useBackupsState from '../../hooks/useBackupsState';
 import useCapabilities from '../../hooks/useCapabilities';
 import useConnection from '../../hooks/useConnection';
+import { STORE_ID } from '../../store';
 import { Backups, Loading as BackupsLoadingPlaceholder } from '../Backups';
+import { useShowBackUpNow } from '../back-up-now/hooks';
+import { BackupNowButton } from '../back-up-now/index';
 import { BackupConnectionScreen } from '../backup-connection-screen';
 import { BackupSecondaryAdminConnectionScreen } from '../backup-connection-screen/secondary-admin';
 import BackupStorageSpace from '../backup-storage-space';
 import ReviewRequest from '../review-request';
-import Header from './header';
 import {
 	useIsFullyConnected,
 	useIsSecondaryAdminNotConnected,
@@ -51,11 +60,30 @@ const Admin = () => {
 
 	const { capabilities, capabilitiesError, capabilitiesLoaded, hasBackupPlan } = useCapabilities();
 
+	const showActivateLicenseLink = useShowActivateLicenseLink();
+	const showBackUpNowButton = useShowBackUpNow();
+
+	const headerActions = useMemo( () => {
+		const items = [];
+		if ( showActivateLicenseLink ) {
+			items.push( <ActivateLicenseLink key="license" /> );
+		}
+		if ( showBackUpNowButton ) {
+			items.push(
+				<BackupNowButton
+					key="backup"
+					variant="primary"
+					tracksEventName="jetpack_backup_plugin_backup_now"
+				>
+					{ __( 'Back up now', 'jetpack-backup-pkg' ) }
+				</BackupNowButton>
+			);
+		}
+		return items.length > 0 ? items : null;
+	}, [ showActivateLicenseLink, showBackUpNowButton ] );
+
 	// If the user is a secondary admin not connected and the site has a backup product,
 	// let's show the login screen.
-	// @TODO: Review the use case where the site is fully connected but the backup product expires and
-	// a secondary admin is not connected. Currently it will display the default `Site backups are
-	// managed by the owner of this site's Jetpack connection.` message.
 	if ( secondaryAdminNotConnected ) {
 		if ( isLoadingBackupProduct ) {
 			return (
@@ -76,10 +104,14 @@ const Admin = () => {
 
 	return (
 		<AdminPage
-			showHeader
 			showFooter
 			moduleName={ __( 'VaultPress Backup', 'jetpack-backup-pkg' ) }
-			header={ <Header /> }
+			title={ __( 'Backup', 'jetpack-backup-pkg' ) }
+			subTitle={ __(
+				'Save changes and restore quickly with one-click recovery.',
+				'jetpack-backup-pkg'
+			) }
+			actions={ headerActions }
 			useInternalLinks={ shouldUseInternalLinks() }
 		>
 			<div id="jetpack-backup-admin-container" className="jp-content">
@@ -388,5 +420,49 @@ const SecondaryAdminConnectionLayout = ( { children } ) => (
 		</Container>
 	</AdminPage>
 );
+
+const useShowActivateLicenseLink = () => {
+	const connectionStatus = useConnection();
+	const isFullyConnected = useIsFullyConnected();
+	const { capabilitiesLoaded, hasBackupPlan } = useCapabilities();
+
+	return useMemo( () => {
+		const connectionLoaded = Object.keys( connectionStatus ).length > 0;
+		if ( ! connectionLoaded ) {
+			return false;
+		}
+
+		if ( ! isFullyConnected ) {
+			return true;
+		}
+
+		if ( ! capabilitiesLoaded ) {
+			return false;
+		}
+
+		return ! hasBackupPlan;
+	}, [ connectionStatus, isFullyConnected, hasBackupPlan, capabilitiesLoaded ] );
+};
+
+const ActivateLicenseLink = () => {
+	const activateLicenseUrl = useSelect( select => {
+		const wpAdminUrl = select( STORE_ID ).getSiteData().adminUrl;
+		return `${ wpAdminUrl }admin.php?page=my-jetpack#/add-license`;
+	}, [] );
+
+	return (
+		<p>
+			{ createInterpolateElement(
+				__(
+					'Already have an existing plan or license key? <a>Click here to get started</a>',
+					'jetpack-backup-pkg'
+				),
+				{
+					a: <a href={ activateLicenseUrl } />,
+				}
+			) }
+		</p>
+	);
+};
 
 export default Admin;
